ml/requirements.txt
fastapi~=0.115.6
torch~=2.5.1
pdf2image~=1.17.0
pillow~=10.4.0
qdrant-client~=1.12.1
colpali_engine~=0.3.4
qwen_vl_utils~=0.0.8
asyncio==3.4.3
transformers==4.46.1